created endpoint to get all parties created
